+++ src/main/java/com/kakaotech/team14backend/post/application/UpdatePostViewCount.java
@@ -2,8 +2,8 @@ package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.common.RedisKey;
 import com.kakaotech.team14backend.common.ScanRedisKey;
-import com.kakaotech.team14backend.exception.PostNotFoundException;
 import com.kakaotech.team14backend.post.domain.Post;
+import com.kakaotech.team14backend.post.exception.PostNotFoundException;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
